Support absolute and relative times in tools
Migrated ParseDates from detections namespace to util. Updated call sites.

First try to parse absolute dates, if that fails, parse as relative. We're watching for various separators as well as 24-hour and 12-hour formats with AM/PM.
