+++ util/dates_test.go
@@ -260,3 +260,52 @@ func TestUnitToDuration(t *testing.T) {
 		})
 	}
 }
+
+func TestParseDate(t *testing.T) {
+	formats := []string{"2006-01-02", "2006/01/02", "2006_01_02"}
+
+	tests := []struct {
+		Name        string
+		Date        string
+		ExpOutput   string
+		ShouldError bool
+	}{
+		{
+			Name:      "Valid date w/Slashes",
+			Date:      "2021/01/01",
+			ExpOutput: "2021-01-01 00:00:00 +0000 UTC",
+		},
+		{
+			Name:      "Valid date w/Dashes",
+			Date:      "2023-10-14",
+			ExpOutput: "2023-10-14 00:00:00 +0000 UTC",
+		},
+		{
+			Name:      "Valid date w/Underscores",
+			Date:      "2022_05_30",
+			ExpOutput: "2022-05-30 00:00:00 +0000 UTC",
+		},
+		{
+			Name:        "Invalid date",
+			Date:        "2022-05-30T12:00:00",
+			ShouldError: true,
+		},
+	}
+
+	for _, test := range tests {
+		test := test
+		t.Run(test.Name, func(t *testing.T) {
+			t.Parallel()
+
+			output, err := ParseDate(test.Date, formats)
+
+			if test.ShouldError {
+				assert.Error(t, err)
+				assert.Contains(t, err.Error(), "unable to parse date string")
+			} else {
+				assert.NoError(t, err)
+				assert.Equal(t, test.ExpOutput, output.String())
+			}
+		})
+	}
+}
